Add mechanism for supporting limits in parameter generation.
* New method areParamsValid for affineEnsemble that tests whether
  a set of parameters are valid.  This defaults to returning true,
  but can be overloaded so that particular problems can have limits
  of various sorts.
* Move step generation out of affineChainSet and into affineEnsemble,
  which is cleaner.
* areParamsValid is used during the stretch move to reject bad steps
  during step generation.  This should be better than just returning
  bad likelihood values.
* affineChainSet::getLastStep now is able to search into previous chunks,
  in case some empty ones have been added on the end.

+++ src/proposedStep.cc
@@ -57,7 +57,7 @@ void proposedStep::sendSelf(MPI_Comm comm, int dest) const {
 	   mcmc_affine::PSTSENDOLIKE, comm);
   MPI_Send(const_cast<double*>(&newLogLike), 1, MPI_DOUBLE, dest, 
 	   mcmc_affine::PSTSENDNLIKE, comm);
-  MPI_Send(const_cast<double*>(&z), 1, MPI_DOUBLE, dest, 
+  MPI_Send(const_cast<float*>(&z), 1, MPI_FLOAT, dest, 
 	   mcmc_affine::PSTSENDZ, comm);
 }
 
@@ -71,7 +71,7 @@ void proposedStep::recieveCopy(MPI_Comm comm, int src) {
 	   comm, &Info);
   MPI_Recv(&newLogLike, 1, MPI_DOUBLE, src, mcmc_affine::PSTSENDNLIKE, 
 	   comm, &Info);
-  MPI_Recv(&z, 1, MPI_DOUBLE, src, mcmc_affine::PSTSENDZ, 
+  MPI_Recv(&z, 1, MPI_FLOAT, src, mcmc_affine::PSTSENDZ, 
 	   comm, &Info);
 }
